hoster/requirements.txt
click==7.1.2
psycopg2-binary==2.8.5
ujson==2.0.3
requests
-e git+https://github.com/mayhem/data-set-hoster.git@first-refactor#egg=datasethoster
